+++ EasyPost.Tests/ServicesTests/ApiKeyServiceTest.cs
@@ -1,4 +1,7 @@
+using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
+using EasyPost.Exceptions.General;
 using EasyPost.Models.API;
 using EasyPost.Tests._Utilities;
 using EasyPost.Tests._Utilities.Attributes;
@@ -37,6 +40,33 @@ namespace EasyPost.Tests.ServicesTests
                 Assert.NotNull(key);
                 Assert.NotNull(key.Key);
             }
+        }
+
+        [Fact]
+        [CrudOperations.Read]
+        public async Task TestRetrieveApiKeys()
+        {
+            UseVCR("retrieve_api_keys");
+
+            User user = await Client.User.RetrieveMe();
+
+            List<ApiKey> apiKeys = await Client.ApiKey.RetrieveApiKeysForUser(user.Id);
+
+            Assert.IsType<List<ApiKey>>(apiKeys);
+        }
+
+        [Fact]
+        [CrudOperations.Read]
+        public async Task TestRetrieveApiKeysChild()
+        {
+            UseVCR("retrieve_api_keys_child");
+
+            const string fakeChildId = "user_123456789";
+
+            // Test suite user has no child users, so this should throw a FilteringError
+            Exception? possibleException = await Record.ExceptionAsync(async () => await Client.ApiKey.RetrieveApiKeysForUser(fakeChildId));
+
+            Assert.IsType<FilteringError>(possibleException);
         }
 
         #endregion